int->size_t

+++ src/Vector.h
@@ -20,7 +20,7 @@ class Vector
 {
 public:
   Vector();
-  Vector(int max_size);
+  Vector(size_t max_size);
   template <size_t MAX_SIZE>
   Vector(T (&values)[MAX_SIZE],
     size_t size=0);
